Add an admin server endpoint to export all stats
While investigating how to export Proxygen related
stats, I found out that we can easily export all
stats to admin server. This is better as there are
very few counters related to Proxygen in HHVM (3 total)
and callers can then easily filter for the stats they
want.

I haven't checked how expensive this is in prod
where we may have concurrent threads trying to
update the counters that we are locking for dumping.
More tests are needed for this.