Collect connection stats for Proxygen
This PR leverages HTTPTransactionTransportCallback
to gather egress/ingress bytes and latencies.

The Proxygen code uses 2 transactions: one for
the incoming request (m_clientTxn) and another one
for the outgoing response (m_pushTxn). For simplicity,
I used a single callback tied to the Transport. I don't
think this is correct for multiplexed protocols (like
HTTP/2.0) but I am not convinced the code handles those
correctly.

One follow-up is to separate (or filter out) stats from
the internal AdminServer that is bundled with external
data right now.

BUG=WIN-2292