Revert "use find_package(c-ares) instead of FindCares.cmake"
This reverts commit 5d1815aec7a8a69c4152dc06acc07a0cd4152a3f.
libc-ares-dev in Ubuntu Jammy is an older version and does
not yet have a CMake package configuration file.
So, keep using our own module for now. Do this as a direct
revert here, since Ubuntu Noble and newer ship a recent enough
version of c-ares to have CMake config files.