WIN-4497 | Override search paths to point to vendored dependencies (#97)
In #96, we established that the mechanism used to replace the glog
and gflags system dependencies with their vendored equivalents in Meta first-party
dependencies wasn't working. Since the vendored version of glog is ABI-incompatible
with the version of glog bundled by Jammy, we need to provide a solution to this problem.

There are three main ways through which glog/gflags are consumed by Meta
first-party dependencies:
*  Explicit `find_package(...)` calls and use of the resulting variables
   to configure linked libraries and include directories.
   Most first-party deps do this.
*  Directly setting an additional include path and produce a staticlib.
   This is generally the case for dependencies where HHVM provides an
   ad-hoc listfile, like squangle or watchmanclient.
*  Transitive dependency on glog/gflags via a dependency on the Folly::folly target.

To handle case 1, we must override the library search paths used by `find_package`
to find our vendored library. The find module for glog is a typical heuristic find
module that calls `find_library` directly. We already pass `GLOG_LIBRARYDIR` as a
`PATHS` argument here, but per the upstream documentation,[1] this path will be considered
only *after* default system library paths have been searched, so if glog/gflags was already
installed on the system, it will prefer the system version. To put our vendored version
first, we must supply an appropriate `<PackageName>_ROOT` override, but this is gated
behind an explicit opt-in into the CMP0074 policy if the project's required CMake version < 3.12.[2]
The gflags case is simpler as the gflags find module is just a wrapper around the gflags CMake config,
so we just need to set `gflags_DIR`.

To handle case 2, we must simply ensure that our ad-hoc listfiles supply an appropriate include directory
for these dependencies.

To handle case 3, we must ensure that folly-deps.cmake links glog and gflags in the right order,
since link order matters for static libraries. It should link gflags after glog, but currently doesn't.

So:
* Set `gflags_DIR` and `Glog_ROOT` for each Meta first-party project referencing them.
* Pass and set a Glog include directory for squangle.
* Flip the order of gflags and glog in folly-deps.cmake to ensure a proper link order.
* Explicitly opt into CMP0074 in all first-party projects that require CMake < 3.12.
  Do this as a direct commit rather than a patch, as we can easily upstream these fixes
  without needing to juggle another patch inside HHVM, except for mvfst which is a submodule.
* Remove seemingly unnecessary patch disabling libunwind in Folly.
  libunwind is required by glog and is already being linked into HHVM.
  Not including it in folly-deps.cmake causes a failure when a first-party project
  tries to produce an executable target while depending on folly (e.g. the thrift compiler).

[1] https://cmake.org/cmake/help/latest/command/find_library.html#command:find_library
[2] https://cmake.org/cmake/help/latest/policy/CMP0074.html