WIN-4497 | Don't pass "separate-code" to the gold linker (#109)
With ENABLE_HARDENING, we try to set `-z separate-code` which is not
understood by the gold linker. So, don't set it if gold is enabled.
Move the ENABLE_LD_GOLD option declaration to Options.cmake alongside
the rest of options for this to work.

The gold linker is officially deprecated and slated for removal by
GNU upstream as of May 2025[1], so a next step would be to remove it
from the build outright. Its main use right now is to use "hot section
ordering" based on a file produced by a run of Meta's oss-performance
WordPress workload eight years ago, which is likely to be different
from webapp and contemporary HHVM.

[1] https://lwn.net/Articles/1007541/