WIN-4504 | Remove notices from Slack ext/json patches (#114)
ext/json contains a set of temporary patches for backwards compatibility
with the jsonc parser, emitting notices when input is encountered that
would fail to parse were it not for these patches.
However, almost every call to `json_decode_mixed` in webapp is conveniently
prefixed with the error silence operator `@`, so we don't really hear from
them much. So, remove the notices outright. If we ever want to invest into
removing the patches, we can implement the logging in a different way.